+++ src/test/java/com/atlassian/db/replica/api/TestStatement.java
@@ -376,4 +376,16 @@ public class TestStatement {
 
         verify(connectionProvider.singleStatement()).setMaxRows(123);
     }
+
+    @Test
+    public void shouldSetMaxFieldSize() throws SQLException {
+        final ConnectionProviderMock connectionProvider = new ConnectionProviderMock();
+        final Connection connection = DualConnection.builder(connectionProvider, new PermanentConsistency()).build();
+        final PreparedStatement statement = connection.prepareStatement(SIMPLE_QUERY);
+
+        statement.setMaxFieldSize(123);
+        statement.executeQuery();
+
+        verify(connectionProvider.singleStatement()).setMaxFieldSize(123);
+    }
 }